+++ ricaun.RevitTest.Console/Revit/Utils/RevitTestUtils.cs
@@ -19,9 +19,6 @@ namespace ricaun.RevitTest.Console.Revit.Utils
     /// </summary>
     public static class RevitTestUtils
     {
-        private const int RevitMinVersionReference = 2021;
-        private const int RevitMaxVersionReference = 2023;
-
         private const int SleepMillisecondsBeforeFinish = 1000;
         private const int SleepMillisecondsDebuggerAttached = 1000;
 
@@ -40,22 +37,17 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                 Log.WriteLine($"RevitTestUtils: {revitVersion}");
                 LoggerTest($"RevitTestUtils: {revitVersion}");
 
-                // Problem with AnavRes.dll / adui22res.dll (version -2020)
-                // Problem with UI (version 2024)
-                var revitVersionMinMax = Math.Min(Math.Max(revitVersion, RevitMinVersionReference), RevitMaxVersionReference);
-                if (RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersionMinMax, out RevitInstallation revitInstallationMinMax))
-                {
-                    LoggerTest($"GetTest Version {revitVersionMinMax}");
-                    Log.WriteLine($"RevitTestUtils: {revitInstallationMinMax.InstallLocation}");
-                    tests = TestEngine.GetTestFullNames(filePath, revitInstallationMinMax.InstallLocation);
-                }
-                else if (RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersion, out RevitInstallation revitInstallation))
+                if (RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersion, out RevitInstallation revitInstallation))
                 {
                     LoggerTest($"GetTest Version {revitVersion}");
                     Log.WriteLine($"RevitTestUtils: {revitInstallation.InstallLocation}");
                     tests = TestEngine.GetTestFullNames(filePath, revitInstallation.InstallLocation);
                 }
 
+                foreach (var exception in TestEngine.Exceptions)
+                {
+                    Log.WriteLine($"RevitTestUtils.Exceptions: {exception}");
+                }
                 if (tests.Length == 0)
                 {
                     throw new Exception($"Unable to read tests class using the Revit version {revitVersion}.");
@@ -111,7 +103,7 @@ namespace ricaun.RevitTest.Console.Revit.Utils
         {
             int timeoutNotBusyCountMax = 10;
 
-            if (timeoutMinutes <= 0) 
+            if (timeoutMinutes <= 0)
                 timeoutMinutes = TimeoutMinutesDefault;
 
             if (revitVersionNumber == 0)
@@ -162,6 +154,7 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                         if (revitInstallation.TryGetProcess(out Process process) == false || forceToOpenNewRevit)
                         {
                             var startRevitLanguageArgument = RevitLanguageUtils.GetArgument(forceLanguageToRevit);
+                            startRevitLanguageArgument = string.Join(" ", startRevitLanguageArgument, Environment.GetEnvironmentVariable("RICAUN_REVITTEST_CONSOLE_PROCESS_ARGUMENTS"));
                             Log.WriteLine($"{revitInstallation}: Start {startRevitLanguageArgument}");
                             process = revitInstallation.Start(startRevitLanguageArgument);
                         }